// src/app/interfaces/memberByTrainerDto.model.ts
export interface MemberByTrainerDto {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  birthDate: string;
  userName: string;
  phoneNumber: string;
  comment: string;
  commentId: number;
  showPriceInput: boolean;
  showInput: boolean;
  anotherInput: number;
  anotherComment: string;
}